Explicitly specifies what user to use with the admin generation script
This change explicitly specifies what user to use with the admin generation script,
so the keycloak would have no issues accessing generated keycloak-add-user.json.

The add-user-keycloak generates keycloak-add-user.json belonging to a user that runs the script.
This means if puppet and keycloak services belong to different users,
without this change there was no guarantee that keycloak could access to this file after creation.